изменили запросы согласно тестовому заданию

+++ internal/models/service.go
@@ -1,9 +1,9 @@
 package models
 
 type Service struct {
-	ID    uint64 `json:"id" db:"id"`
+	ID    int    `json:"id" db:"id"`
 	Name  string `json:"name" db:"name"`
-	Price int64  `json:"price" db:"price"`
+	Price int    `json:"price" db:"price"`
 }
 
 func NewService() Service {
